refactor; move out import context
